+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.3] - 2024-13-03
+### Added
+- GRPC Stream Middleware for metrics
+- GRPC Unary Middleware for metrics
+## Removed
+- GRPC Middleware for metrics (client)
+
 ## [0.0.2] - 2024-13-03
 ### Added
 - JSONRPC Middleware for metrics